twitter-sentiment-app-v1.0.5/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
pydantic==2.8.2